Note all locations of base distros for packages
We indicate that there are two different places we keep a list of
distros, but in fact, there are four.  Let's make a note at each place
indicating these and refer to the others to make sure they're updated.

A good future direction would be to store this data in some sort of data
file and read it in each case, but we don't currently do that.