docs: add manual page for completion command
In commit 15723008062b941b53743295784554b2f3ba875e of PR #5311 the
"git lfs completion" command was introduced, utilizing the support
provided by the spf13/cobra package to generate tab-completion scripts
for a number of shells, including the Bash, fish, and Zsh shells.

However, no manual page was added at the same time, although help text
was included using the spf13/cobra package's format for long command
descriptions.  Unfortunately, as described in an earlier commit in this
PR, the Git LFS client overrides the default help functions defined by
the spf13/cobra package and so there was no way for a user to access
this long-format help text for the "git lfs completion" command.

We therefore removed the original help text for the command, and now
reintroduce a revised and expanded version of the text in the form of
a manual page for the git-lfs-completion(1) command.

The full text of this new manual page attempts to explain the features
and limitations of the current tab-completion scripts, especially the fact
that the Git LFS client does not yet offer suggested lists of terms
specific to Git, such as ref or remote names, but that the tab-completion
scripts for Git LFS should integrate well with the tab-completion
scripts for Git itself, provided the correct one is used for a given
shell.

We also try to include examples of the configuration of tab-completion
for each type of shell and for both a user's current session as well
as for future sessions.  Complete documentation of all such per-shell
configuration options is beyond the scope of the Git LFS project,
of course, but we aim to provide a sufficient overview so as to guide
users to a working solution and forestall support requests as much
as possible.