commands/run.go: remove powershell completion
In commit 15723008062b941b53743295784554b2f3ba875e of PR #5311 the
"git lfs completion" command was introduced, utilizing the support
provided by the spf13/cobra package to generate tab-completion scripts
for a number of shells, including Bash, fish, Zsh, and PowerShell.

For the shells other than PowerShell, the generated completion scripts
are intended to work with the tab-completion scripts for Git commands
that are available from either the Git project (for Bash) or as an
implementation native to the shell.  This coordination permits users
to receive Git LFS command suggestions when they type "git lfs [Tab]",
for instance, instead of only when they type "git-lfs [Tab]".

In the case of PowerShell, Git command tab-completion is provided by a
third-party project (https://github.com/dahlbyk/posh-git), and while
it may eventually be possible to integrate that project's scripts with
the ones generated by the spf13/cobra package, at the present time
the Git LFS core team lacks the time and expertise to develop such
a solution.  Thus the script currently generated for Git LFS command
completion in PowerShell would only provide suggestions when users
enter the "git-lfs" program name, rather than "git lfs".

As we do not want to offer a subpar user experience and one which the
Git LFS core team is unable to properly support, for the present time we
remove the PowerShell option from the "git lfs completion" command.

Note that PowerShell is primarily (but not exclusively) used on Windows,
where most Git LFS users will have installed Git LFS as part of the
Git for Windows project.  Git for Windows also provides an emulation
of the Bash shell, where the Bash tab-completion script generated by
the "git lfs completion" command should function as expected.