commands/run.go: remove completion help text
In commit 15723008062b941b53743295784554b2f3ba875e of PR #5311 the
"git lfs completion" command was introduced, utilizing the support
provided by the spf13/cobra package to generate tab-completion scripts
for a number of shells.  Short and long help text was provided at
the same time, in the formats offered by the spf13/cobra package.

However, the Git LFS client explicitly overrides the default help
functions of the spf13/cobra package, and replacing them with functions
which output versions of the Git LFS manual pages that are pre-rendered
by our build process.  Thus the long-format help text included in
PR #5311 can not be viewed by end users.

As this text is also somewhat terse, having been copied from the
example text provided by the spf13/cobra package, we remove it now
with the expectation that in a subsequent commit in this PR we will
provide a more comprehensive set of instructions in a new manual page
for the "git lfs completion" command.

We also remove the short help text for the "git lfs completion" command,
as the majority of other Git LFS commands (with the exception of the
"git lfs help" command) do not define these text strings, and in a
subsequent commit in this PR we will suppress their inclusion in the
command lists output by the completion scripts, so as not to make
the formatting of those lists irregular and uneven.