commands/run.go: alphabetize completion shell opts
In commit 15723008062b941b53743295784554b2f3ba875e of PR #5311 the
"git lfs completion" command was introduced, utilizing the support
provided by the spf13/cobra package to generate tab-completion scripts
for a number of shells, including the Bash, fish, and Zsh shells.

We now reorder the list of these shells and their relevant case switches
so they are in alphabetical order.