commands/run.go: use v2 bash completion script
In commit 15723008062b941b53743295784554b2f3ba875e of PR #5311 the
"git lfs completion" command was introduced, utilizing the support
provided by the spf13/cobra package to generate tab-completion scripts
for a number of shells, including the Bash shell.  The completion
script for the Bash shell is currently generated using the spf13/cobra
package's legacy implementation, which is returned by the
(*Command).GenBashCompletion() method.

However, the spf13/cobra package's maintainers advise using their
newer implementation of tab-completion for the Bash shell, which is
returned by the (*Command).GenBashCompletionV2() method, as this
version of the script is shorter and aligned with the tab-completion
scripts now generated for other shells.

In particular, the new version of the Bash tab-completion script has
support for the program to return custom, dynamic completion lists using
per-command functions.  While the current Git LFS client does not
implement any of these ValidArgsFunction() or RegisterFlagCompletionFunc()
functions, we anticipate doing so in the future so as to be able to
provide lists of Git-specific values such as ref and remote names,
when appropriate for a particular Git LFS command.

Note that in order for the new Bash completion script for Git LFS to
integrate properly with the Bash completion script from the Git project,
we have to revise the script generated by the spf13/cobra package so it
rewrites the set of command words when the user types "git lfs ...[Tab]"
such that the initial two words are replaced with the single word
"git-lfs".  This requires a more substantial replacement of a portion
of the generated script than was the case with the legacy implementation.