commands/run.go: tidy zsh completion script search
In a prior commit in this PR we revised the "git lfs completion" command
so it would generate a tab-completion script for the Bash shell using
the newer implementation offered by the spf13/cobra package's
(*Command).GenBashCompletionV2() method, and in doing so also added
a search-and-replace operation to revise the generated script such that
it would integrate with Git's own tab-completion script for Bash.

We already perform a similar search-and-replace operation to revise
the script generated for the Zsh shell by the spf13/cobra package's
(*Command).GenZshCompletion() method.  These script revisions were
introduced in commit 4c9987de9be07bd760e01b39eb9cd72d155c1b25 of PR #5311
so that the tab-completion script for Git LFS would integrate with the
Zsh shell's native tab-completion for Git.

We now refactor the script generation and modification for Zsh to more
closely align with that used for the Bash shell, which should also
slightly improve the clarity of the code and make its purpose more
obvious.