push: allow empty input in --stdin mode
When we're scripting, we may want to process any number of object IDs in
a pipeline and then pipe them to `git lfs push --object-id --stdin`.
However, right now we fail if there are zero object IDs, which means
that the user must explicitly check for this case and skip the attempt
to push, which makes pipelines and scripting much more complicated.

To simplify the behaviour of scripts, let's allow an empty input with
`--stdin` so that users don't have to adjust their scripts for this
case.  If we receive no object IDs on standard input, then we simply do
nothing.