push: don't allow empty list of refs
Right now, a user can invoke `git lfs push` with a remote and no refs.
Surprisingly for some users, this pushes nothing, since no refs were
given.

To help users invoke the command in a useful way, if there are no
arguments on the command line and `--all` was not specified, then
complain, unless we're reading from standard input.  The reason we allow
empty lists of arguments with standard input is that this makes
scripting easier by not forcing scripts to specifically consider
whether the input is empty.