script/packagecloud: stop printing Markdown links
We used to print these Markdown links because we were expecting to copy
and paste them into the release data.  Since the release process is now
heavily automated, we no longer need to do that, so printing this data
is now no longer useful.  Remove it so that we don't have to maintain it
anymore.