run_dockers: use distro-tool for finding distros
Now that we've centralized our distro information in one place, call our
distro tool to acquire the list of images to build.  To make it easier
to handle this case by reading the output of a command and splitting on
whitespace, let's switch from a bash array to a standard shell string.