script/upload: use distro-tool to generate list of PackageCloud links
Now that we have a single location for all of our distro information,
use the distro tool, which can extract this information, to emit the
links to the packages.  Remove the now useless comment.

Note that the removed text and the text output by the command
substitution are identical according to diff, so this should result in
no functional change.