FAQ: add an entry about proxies
Lots of people have questions about how to use proxies with Git LFS.
Document that they can be used in the same way as with Git, except that
we only support Basic authentication.  Tell people how to format the
value.

Finally, indicate to users that their proxy must not modify the data in
any way and that we don't encourage the use of TLS MITM proxies.  As has
been documented in multiple scientific papers, including those below,
these have been shown to be the cause of substantial security issues:

* https://arxiv.org/abs/1809.08729
* https://jhalderm.com/pub/papers/interception-ndss17.pdf