mangen: allow a second level of header in help output
In manual pages, it can be helpful to add a second level of sections.
Let's teach our help output how to add handle that by underscoring the
name of the section with tildes.  This is the approach that AsciiDoc
itself uses for underlined headers.

Note that traditionally, on Linux (and probably BSD) systems, this level
of headers is written in mixed case, not all-uppercase, as can be seen
in groff(1), so don't try to mangle the case of those headers.