docs/howto: use generic numbering in release steps
Our current documentation of our steps to release a new Git LFS version
uses example version numbers which start with major version 2.  As we
have since released v3.0 and anticipate a potential v4.0 in the future,
we can make our documentation more generic by using a "vM.N.P" format
in all the notes and examples.

We also add parentheses when referring to a prior release to help visually
distinguish the segments of the semantic versioning format in these
examples; for instance, "vM.(N-1).0" instead of "vM.N-1.0".