docs/howto: update expected release timescales
In our documentation of the steps to release a new version of Git LFS,
we provide descriptions of the different types of versions, according
to the semantic versioning scheme.  In these descriptions we offer
suggested time scales for the different version types, such as that
major releases will be done every one to two years.  These suggestions
date from the era when the documentation was first written, in 2018.

As Git LFS is now a more mature project and the release schedule is
less frequent than it was in the past, we adjust the suggested time
scales for different version types to more accurately reflect our
current development and release process.