docs/howto: drop minor release branch instructions
In our documentation of the steps to release a new version of Git LFS,
we have a section detailing the process for a minor version release
which states that we create a new "release-M.N" branch for every
minor release.

However, this does not accord with our current practice, where we only
create such a branch when we need to issue a patch release.  Before
updating the instructions for creating such a patch release to explain
this step, we first just remove the section pertaining to minor releases,
since there are no extra steps for a minor release compared to a major one.