docs/howto: update patch release instructions
In a previous commit in this PR, we removed our legacy instructions
pertaining to the creation of a "release-M.N" branch immediately
following the release of a new minor version of Git LFS, as we no
longer create such a branch in all cases, but only when we need
to create a subsequent patch release.

We now update our instructions for building a patch release to
include a description of when and how to create a "release-M.N" branch
onto which we then cherry-pick commits from the mainline.

We also edit our instructions for performing the cherry-pick operations
to explain more clearly how we identify the mainline parent of each
merge commit, and make some other minor improvements to our instructions
as well.