docs/howto: use generic examples in release steps
Our documentation of the steps to release a new version of Git LFS
have several examples of specific command-line actions or file changes
which use actual legacy commits for their sample data.  As these are
not particularly current anymore (for instance, they refer to the
previous major version), we revise them to be more generic, with
placeholders for any specific version numbers, GitHub PR numbers, etc.

As well, we update the "git tag" command in one example to include
a message with the -m option, as that is expected by our current
release process, and we add a "git show" command to illustrate what
the signed tag should look like after it has been created.  We also
use a simpified format for the Git commands in the examples so that
their output is more compact, making the examples more generic.

Finally, we update the hostname of the Git LFS project home page to
git-lfs.com from git-lfs.github.com, as that change has been made
since these instructions were first written.