docs/howto: add new Homebrew update release step
The final step in our instructions for releasing a new Git LFS version
advise the maintainer to update our fork of the Homebrew/homebrew-core
GitHub repository.  However, we recently archived this repository as
it is no longer applicable to the release process, and is not needed
to provide an updated Homebrew formula for Git LFS.

Instead, we add a new step to our release instructions which explains
how to run the "brew bump-formula-pr" command on a macOS system to
automatically create a GitHub PR for the Homebrew project that will
advance the version number in the git-lfs Homebrew formula to that of
the latest Git LFS release.