docs/howto: edit release changelog creation steps
In our documentation of the steps to release a new version of Git LFS,
the first step pertains to the creation of release notes and an updated
changelog on a new "release-next" branch.  This step begins with some
description of our semantic versioning scheme, specifically relating
to minor versions, and then the following bullet points include an
explanation of the use of our script/changelog script to help populate
the CHANGELOG.md file, with specific example commands for a new minor
version only.

In order to provide more comprehensive guidance, we update this first
step to describe all three types of versions (major, minor, and patch),
and provide examples of how to run the script/changelog script for
each type.

We also expand our documentation of the need to carry forward the changelog
from a patch release series into the changelog for the next major or
minor release, as we currently only discuss minor releases in this regard.

Finally, we note that one can use the GITHUB_TOKEN environment variable
instead of entering GitHub credentials into a ~/.netrc file, and
that the debian/changelog file needs to be manually updated after
running the script/update-version script.