docs/howto: update release PR instructions
In our documentation of the process by which we release new versions
of Git LFS, the first step describes how a new "release-next" branch
should be prepared, and the second step describes how to create a
GitHub pull request from that branch, check the status of its CI jobs,
and announce the forthcoming release to interested parties using the
new PR.

One detail missing from the current set of instructions in this second
step is the fact that our GitHub Actions release workflow is not
triggered by a new PR, and so while the PR will cause the regular CI
workflow to run, it will not do so for the release workflow, which is
a crucial part of the later steps.  Because the release workflow can
accumulate regressions and bugs between releases, we have sometimes
encountered problems during release where it fails and we need to
immediately issue a patch release after fixing the problems.  Hence,
it is advisable to try to run the release workflow (as much as possible)
in a dry-run operation beforehand.

We therefore add some bullet points to our second release step on how
to do this safely, without actually uploading release assets to
Packagecloud.

We also add bullet points which mention adding a "release" label to
the PR, and uploading a set of pre-built binaries for testing purposes
by interested parties.  As well, we update several other points and
reformat the instruction for greater clarity.