docs/howto: add release upload script dependencies
The third step in our description of the process of releasing a new
Git LFS version advises the use of our script/upload script, which
uploads release artifacts and checksum files to GitHub, after generating
GPG signatures for the checksum files using the user's PGP key.

This script actually runs the script/hash-files as well as other
scripts and tools, and collectively these have a range of requirements,
such as a modern Ruby installation with the OpenSSL gem, GnuPG, etc.,
so we add a paragraph to our documentation of this third step which
mentions these dependencies.