Make track handle backslashes correctly on Unix
On Windows, backslashes are path separators and we need to rewrite them
into forward slashes if we want Git to honour the path provided.
However, on Unix, backslashes are ordinary characters and can be used
anywhere in a path.

To make `git lfs track --filename` work correctly for paths with
backslashes on Unix, we must avoid converting them to forward slashes,
and we need to unescape them when unescaping the paths.  Note that we
must escape _before_ all other escapes and unescape _after_ all other
unescaping so that we don't accidentally try to re-escape a pattern that
we've just escaped (or vice-versa).

In addition, this highlights an additional subtle bug we've failed to
notice so far: the `core.quotepaths` variable we've used in our
invocation to `git ls-files` doesn't prevent all escaping; instead, it
only prevents escaping of high bytes, which doesn't help us with
sequences that contain backslashes or newlines.  To handle those
gracefully, we need the `-z` option, which uses NUL-terminated
filenames.

Update the tests for our new backslash behaviour on Unix and add new
tests that we gracefully handle backslashes on Unix.  Check specifically
for the error we would get if `git ls-files` produced unexpected output
and fail if so.