git: use `git var` to find system gitattributes file
At the moment, Git doesn't provide us a way to know where the system
gitattributes file is.  To help find it, we use the `PREFIX` environment
variable and tack on `etc/gitattributes`.  However, as a practical
matter, nobody sets `PREFIX`, so typically we always use
`/etc/gitattributes`.

That being said, Git has just learned a way to provide this information
using `git var GIT_ATTR_SYSTEM` to print this information, and this will
be included in Git 2.42.  If we have such a version of Git, use it to
find this information.  If it fails, return an error, but if it simply
prints no values (such as with `GIT_ATTR_NOSYSTEM`), then that means Git
does not consider us to have a system gitattributes file, and in that
case, ignore trying to load that data.