lfsapi: avoid indexing a nonexistent value
In some cases, we might end up with an entry that's missing either a
username or password, and in such a case, we wouldn't want to panic,
like we do now.  Solve this by using the function `FirstEntryForKey` to
look up the value and return an empty string if none is present.