t: add a test helper to print the mtime of a file
In a future commit, we'll want to verify that the mtime of a file isn't
changed.  However, because `ls` doesn't provide a great deal of
granularity, we'll need a test helper to print the number of nanoseconds
involved.

On Linux, we could use `find` with `-printf` or `stat`, but this won't
work on macOS, since this option isn't available in the BSD version of
`find` and `stat` won't exist.  Thus, a test helper is the simplest
approach, so implement one.