lfs: don't update mtime on empty files on pull
When a user performs a `git lfs pull`, we update all of the files in the
working tree that are pointers (and aren't excluded) to have their
correct content.  However, if the files are empty, they don't need a
change, and some users find the fact that we update the mtime of the
file to be undesirable.

Let's avoid this by explicitly checking to see if the file is empty and
if it's supposed to be and not updating it.