docker: remove unused scripts
When the scripts to build Docker images of Git LFS were first introduced
in PRs #511 and #555, an optional capabililty was included to sign the RPM
or Debian packages after building them in the containers.  This option
was triggered by the presence of a git-lfs_*.key file alongside the
Dockerfile for a given OS and version.

However, we have not published Docker containers to the Docker Hub or
any other registry for at least five years, and we do not use or maintain
this signing capability, so we can remove the associated script code.
If we ever decide to begin publishing containers again, we can restore
this functionality from our Git history.

The docker/test_dockers.bsh script is also out of date and unused, so we
remove it as well.