Limit CopyCallbackFile to print every 250 ms
Prior to this commit, the callback would print every single iteration of io.Copy.
This results in a print every 32kb and
can slow clean operations by several orders of magnitude
when GIT_LFS_PROGRESS is set.
By limiting the print to every quarter second in this PR,
most of negative impact is mitigated.