docs/linux-build.md: remove outdated documentation
The last substantive update to the docs/linux-build.md file was made
in PR #480 in 2015, and the information in the file is no longer
accurate or relevant.

While the `rpm/build_rpms.bsh` script mentioned in the text still exists,
we typically run that script only within Docker containers, and we have
updated our Docker build documentation both in prior commits in this PR
and in PR git-lfs/build-dockers#57.  Therefore we simply remove the old
docs/linux-build.md file in favour of our revised documentation, which
describes how to build and use Docker images to generate both Debian and
RPM packages of Git LFS.