{t/}Makefile: force use of local Go toolchain
As of Go version 1.21, the "go" command may, under certain circumstances,
fetch a different version of the Go toolchain than is provided locally, as
described in #5477, PR git-lfs/build-dockers#58, and the Go documentation:

https://go.dev/doc/toolchain

In general, we do not expect this to occur, as we specify an older
Go version in our go.mod file and so the locally-installed Go toolchain
should be newer than that version and therefore sufficient.  We also do
not run any of the commands, such as "go get", "go install", and "go run",
which might cause Go to attempt to fetch a newer version of itself.

Nevertheless, to be cautious and ensure we only try to run with the
Go version we have locally installed, we add the GOTOOLCHAIN=local
environment variable to our Makefile.  Defining this variable will
guarantee that Go will stop with an error rather than proceed to
download and run another version of itself.