t/Makefile: use variable to run Go commands
The Makefile in our "t" test directory was introduced in commit
4a380ce0090deb0714832e66bdc38af92b25e2a9 of PR #3125, partly
so as to provide a convenient way to build the test utilities
in the t/cmd directory, which are written in Go.

While the test Makefile sets a GO variable like the one used throughout
our primary Makefile, it does not make use of it and instead runs
"go build" directly.  We revise the "go build" commands now to use
the GO variable in order to bring them into alignment with our primary
Makefile.