t/t-unlock.sh: test unlock by path with refspec
In PR #2773 the option to include a Git reference specification in
Git LFS locking API requests was introduced, and several tests were
added to the t/t-unlock.sh test suite and others modified so as to
help validate this behaviour.  As well, the lfstest-gitserver.go
test server was altered to check for an expected "refspec" query parameter
in lock listing requests, but only if the test repository name had the
trailing suffix "branch-required".  This permitted the test server to
mimic existing production services whose APIs did not expect or check
the "refspec" query parameter, but also emulate those services which
chose to fully implement the new API feature and only list locks
matching the given "refspec".

Four tests of the "git lfs unlock" command were added or changed in
order to check the command's behaviour when support for the "refspec"
argument was added in PR #2773.  When the --id option is not used
with this command, it makes an initial GET request to the Git LFS
locking API to retrieve the lock's ID, if any, based on the file
path provided to the command.  With the changes in the PR, the current
"refspec" was also sent to the API to be used in the lookup operation.

In commit 257df970f07900bd1a6b79619ee553afce674b33, the "unlocking a
lock by path" test was revised into the "unlocking a lock by path with
good ref" test, and the "unlocking a lock by path with tracked ref",
"unlocking a lock by path with bad ref", and "unlocking a lock by id
with bad ref" tests were added.  All of these tests used test repository
names with the "branch-required" suffix so as to cause the test API
service to require a "refspec" query parameter in client GET requests,
and check that its value matched the branch name extracted from the
test repository name as the word which preceded the suffix (e.g., "main"
in the test repository name "unlock-by-path-main-branch-required").

But then in commit 9027b941e8ec9e053324a73a14d1bc362009f9b8 of the
same PR, the three "unlocking a lock by path*" tests were all changed
to actually provide a lock ID with the --id argument to the "git lfs
unlock" command.  The test names and test repository names were not
updated, though.  This change meant that the "unlocking a lock by path
with bad ref" test was identical to the "unlocking a lock by id with
bad ref" test, except for its names; it also meant that no tests
validated the command's behaviour when only a file path was provided
on the command line, not a lock ID, and the API service was configured
to require a "refspec".

The consequence of this omission was that no tests detected the fact
that the "git lfs unlock" command did not send a "refspec" argument
to the API service even when attempting to look up a lock by file path.

As this oversight in the command's implementation has now been corrected
in a prior commit in this PR, we want to ensure that our test suite
validates that the command operates successfully with both older API
services, which do not filter on "refspec" query parameters when searching
for locks, and those which do.

As a first step, we change the "unlocking a lock by path with good ref"
and "unlocking a lock by path with tracked ref" tests to actually use a
file path rather than an ID in their "git lfs unlock" commands, while
replicating them as "unlocking a lock by id*" tests which continue to
use the --id option.  We also alter the "unlocking a lock by path
with bad ref" test to actually use a file path as well.  With these
changes we now have a set of tests which confirm that the "git lfs
unlock" command, when supplied with a file path and not a lock ID,
sends a "refspec" query parameter in its lock listing request, and
either succeeds or fails depending on whether the "refspec" matches
that of the lock recorded by the API service.

Note that without the fix to the lockIdFromPath() method of the Client
structure in the "locking" package from the prior commit in this PR,
our new "unlocking a lock by path with good ref", "unlocking a lock by
path with tracked ref", and "unlocking a lock by path with bad ref"
tests all fail.  Previously the tests with these names all succeeded
despite the problem in the "git lfs unlock" command because the tests
all used the command with the --id option and so did not exercise the
incorrect implementation of the file path option.